# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Counselor Backend API
  version: "0.1.0"
  description: |
    Web-only MVP backend for private reflection, partner-aligned insights, screenings, and gentle escalation.
servers:
  - url: https://api.example.com
paths:
...
  /consents:
    get:
      summary: List consents
      parameters:
        - in: query
          name: user_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Consents listed }
  /consents/grant:
    post:
      summary: Grant consent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string, format: uuid }
                scope: { type: string }
                target_partner_id: { type: string, format: uuid }
              required: [user_id, scope]
      responses:
        "200": { description: Consent granted }
  /consents/revoke:
    post:
      summary: Revoke consent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string, format: uuid }
                scope: { type: string }
                target_partner_id: { type: string, format: uuid }
              required: [user_id, scope]
      responses:
        "200": { description: Consent revoked }
...